feat: Improve platform setup (#51)
* feat: Move platforms into folders

* feat: Dont rely on fs.edDir to find platform classes

* feat: Use static OAuth2Service

* feat: Move all apis into separate files

* chore: Add docblocks

* chore: Tested and working ..

.. except, twitter and reddit keys are not
automatically refreshed anymore. will solve this
differently.

---------
