+++ src/platforms/Ayrshare/AsInstagram.ts
@@ -1,7 +1,7 @@
 import * as path from "path";
 import * as sharp from "sharp";
 
-import Ayrshare from "../Ayrshare";
+import Ayrshare from "./Ayrshare";
 import Folder from "../../models/Folder";
 import Logger from "../../services/Logger";
 import { PlatformId } from "..";
