+++ src/platforms/Ayrshare/AsTikTok.ts
@@ -1,4 +1,4 @@
-import Ayrshare from "../Ayrshare";
+import Ayrshare from "./Ayrshare";
 import Folder from "../../models/Folder";
 import { PlatformId } from "..";
 import Post from "../../models/Post";
