+++ src/platforms/Ayrshare/Ayrshare.ts
@@ -1,13 +1,13 @@
 import * as fs from "fs";
 import * as path from "path";
 
-import Folder from "../models/Folder";
-import Logger from "../services/Logger";
-import Platform from "../models/Platform";
-import { PlatformId } from ".";
-import Post from "../models/Post";
-import { PostStatus } from "../models/Post";
-import Storage from "../services/Storage";
+import Folder from "../../models/Folder";
+import Logger from "../../services/Logger";
+import Platform from "../../models/Platform";
+import { PlatformId } from "..";
+import Post from "../../models/Post";
+import { PostStatus } from "../../models/Post";
+import Storage from "../../services/Storage";
 import { randomUUID } from "crypto";
 
 /**
