+++ src/platforms/Instagram/Instagram.ts
@@ -2,14 +2,14 @@ import * as fs from "fs";
 import * as path from "path";
 import * as sharp from "sharp";
 
-import Folder from "../models/Folder";
-import InstagramAuth from "../auth/InstagramAuth";
-import Logger from "../services/Logger";
-import Platform from "../models/Platform";
-import { PlatformId } from ".";
-import Post from "../models/Post";
-import { PostStatus } from "../models/Post";
-import Storage from "../services/Storage";
+import Folder from "../../models/Folder";
+import InstagramApi from "./InstagramApi";
+import InstagramAuth from "./InstagramAuth";
+import Logger from "../../services/Logger";
+import Platform from "../../models/Platform";
+import { PlatformId } from "..";
+import Post from "../../models/Post";
+import { PostStatus } from "../../models/Post";
 
 /**
  * Instagram: support for instagram platform.
@@ -20,21 +20,23 @@ import Storage from "../services/Storage";
  */
 export default class Instagram extends Platform {
   id: PlatformId = PlatformId.INSTAGRAM;
-  GRAPH_API_VERSION: string = "v18.0";
+  api: InstagramApi;
+  auth: InstagramAuth;
 
   constructor() {
     super();
+    this.auth = new InstagramAuth();
+    this.api = new InstagramApi();
   }
 
   /** @inheritdoc */
   async setup() {
-    const auth = new InstagramAuth();
-    await auth.setup();
+    await this.auth.setup();
   }
 
   /** @inheritdoc */
   async test() {
-    return this.get("me");
+    return this.api.get("me");
   }
 
   /** @inheritdoc */
@@ -82,7 +84,7 @@ export default class Instagram extends Platform {
 
   /** @inheritdoc */
   async publishPost(post: Post, dryrun: boolean = false): Promise<boolean> {
-    Logger.trace("Instagram.publishPost", post, dryrun);
+    Logger.trace("Instagram.publishPost", post.id, dryrun);
 
     let response = dryrun ? { id: "-99" } : ({} as { id: string });
     let error = undefined;
@@ -144,7 +146,7 @@ export default class Instagram extends Platform {
   ): Promise<{ id: string }> {
     const photoId = (await this.fbUploadPhoto(file))["id"];
     const photoLink = await this.fbGetPhotoLink(photoId);
-    const container = (await this.postJson("%USER%/media", {
+    const container = (await this.api.postJson("%USER%/media", {
       image_url: photoLink,
       caption: caption,
     })) as { id: string };
@@ -155,7 +157,7 @@ export default class Instagram extends Platform {
     if (!dryrun) {
       // wait for upload ?
       // https://github.com/fbsamples/reels_publishing_apis/blob/main/insta_reels_publishing_api_sample/utils.js#L23
-      const response = (await this.postJson("%USER%/media_publish", {
+      const response = (await this.api.postJson("%USER%/media_publish", {
         creation_id: container.id,
       })) as { id: string };
       if (!response?.id) {
@@ -184,7 +186,7 @@ export default class Instagram extends Platform {
   ): Promise<{ id: string }> {
     const videoId = (await this.fbUploadVideo(file))["id"];
     const videoLink = await this.fbGetVideoLink(videoId);
-    const container = (await this.postJson("%USER%/media", {
+    const container = (await this.api.postJson("%USER%/media", {
       video_url: videoLink,
       caption: caption,
     })) as { id: string };
@@ -195,7 +197,7 @@ export default class Instagram extends Platform {
     if (!dryrun) {
       // wait for upload ?
       // https://github.com/fbsamples/reels_publishing_apis/blob/main/insta_reels_publishing_api_sample/utils.js#L23
-      const response = (await this.postJson("%USER%/media_publish", {
+      const response = (await this.api.postJson("%USER%/media_publish", {
         creation_id: container.id,
       })) as { id: string };
       if (!response?.id) {
@@ -229,7 +231,7 @@ export default class Instagram extends Platform {
       const videoLink = await this.fbGetVideoLink(videoId);
       uploadIds.push(
         (
-          await this.postJson("%USER%/media", {
+          await this.api.postJson("%USER%/media", {
             is_carousel_item: true,
             video_url: videoLink,
           })
@@ -244,7 +246,7 @@ export default class Instagram extends Platform {
       const photoLink = await this.fbGetPhotoLink(photoId);
       uploadIds.push(
         (
-          await this.postJson("%USER%/media", {
+          await this.api.postJson("%USER%/media", {
             is_carousel_item: true,
             image_url: photoLink,
           })
@@ -253,7 +255,7 @@ export default class Instagram extends Platform {
     }
 
     // create carousel
-    const container = (await this.postJson("%USER%/media", {
+    const container = (await this.api.postJson("%USER%/media", {
       media_type: "CAROUSEL",
       caption: post.body,
       children: uploadIds.join(","),
@@ -266,7 +268,7 @@ export default class Instagram extends Platform {
 
     // publish carousel
     if (!dryrun) {
-      const response = (await this.postJson("%USER%/media_publish", {
+      const response = (await this.api.postJson("%USER%/media_publish", {
         creation_id: container["id"],
       })) as {
         id: string;
@@ -298,7 +300,7 @@ export default class Instagram extends Platform {
     body.set("published", "false");
     body.set("source", blob, path.basename(file));
 
-    const result = (await this.postFormData("%PAGE%/photos", body)) as {
+    const result = (await this.api.postFormData("%PAGE%/photos", body)) as {
       id: "string";
     };
 
@@ -315,7 +317,7 @@ export default class Instagram extends Platform {
    */
   private async fbGetPhotoLink(id: string): Promise<string> {
     // get photo derivatives
-    const photoData = (await this.get(id, {
+    const photoData = (await this.api.get(id, {
       fields: "link,images,picture",
     })) as {
       link: string;
@@ -356,7 +358,7 @@ export default class Instagram extends Platform {
     body.set("published", "false");
     body.set("source", blob, path.basename(file));
 
-    const result = (await this.postFormData("%PAGE%/videos", body)) as {
+    const result = (await this.api.postFormData("%PAGE%/videos", body)) as {
       id: string;
     };
 
@@ -373,7 +375,7 @@ export default class Instagram extends Platform {
    */
 
   private async fbGetVideoLink(id: string): Promise<string> {
-    const videoData = (await this.get(id, {
+    const videoData = (await this.api.get(id, {
       fields: "permalink_url,source",
     })) as {
       permalink_url: string;
@@ -383,156 +385,5 @@ export default class Instagram extends Platform {
       throw Logger.error("No source url found for video " + id);
     }
     return videoData["source"];
-  }
-
-  // API implementation -------------------
-
-  /**
-   * Do a GET request on the graph.
-   * @param endpoint - the path to call
-   * @param query - querystring as object
-   * @returns parsed response
-   */
-
-  private async get(
-    endpoint: string = "%USER%",
-    query: { [key: string]: string } = {},
-  ): Promise<object> {
-    endpoint = endpoint.replace(
-      "%USER%",
-      Storage.get("settings", "INSTAGRAM_USER_ID"),
-    );
-    endpoint = endpoint.replace(
-      "%PAGE%",
-      Storage.get("settings", "INSTAGRAM_PAGE_ID"),
-    );
-
-    const url = new URL("https://graph.facebook.com");
-    url.pathname = this.GRAPH_API_VERSION + "/" + endpoint;
-    url.search = new URLSearchParams(query).toString();
-    const accessToken = Storage.get("auth", "INSTAGRAM_PAGE_ACCESS_TOKEN");
-    Logger.trace("GET", url.href);
-    return await fetch(url, {
-      method: "GET",
-      headers: accessToken
-        ? {
-            Accept: "application/json",
-            Authorization: "Bearer " + accessToken,
-          }
-        : {
-            Accept: "application/json",
-          },
-    })
-      .then((res) => this.handleApiResponse(res))
-      .catch((err) => this.handleApiError(err));
-  }
-
-  /**
-   * Do a Json POST request on the graph.
-   * @param endpoin - the path to call
-   * @param body - body as object
-   * @returns the parsed response as object
-   */
-
-  private async postJson(
-    endpoint: string = "%USER%",
-    body = {},
-  ): Promise<object> {
-    endpoint = endpoint.replace(
-      "%USER%",
-      Storage.get("settings", "INSTAGRAM_USER_ID"),
-    );
-    endpoint = endpoint.replace(
-      "%PAGE%",
-      Storage.get("settings", "INSTAGRAM_PAGE_ID"),
-    );
-
-    const url = new URL("https://graph.facebook.com");
-    url.pathname = this.GRAPH_API_VERSION + "/" + endpoint;
-    Logger.trace("POST", url.href);
-    return await fetch(url, {
-      method: "POST",
-      headers: {
-        Accept: "application/json",
-        "Content-Type": "application/json",
-        Authorization:
-          "Bearer " + Storage.get("auth", "INSTAGRAM_PAGE_ACCESS_TOKEN"),
-      },
-      body: JSON.stringify(body),
-    })
-      .then((res) => this.handleApiResponse(res))
-      .catch((err) => this.handleApiError(err));
-  }
-
-  /**
-   * Do a FormData POST request on the graph.
-   * @param endpoint - the path to call
-   * @param body - body as object
-   * @returns the parsed response as object
-   */
-
-  private async postFormData(
-    endpoint: string,
-    body: FormData,
-  ): Promise<object> {
-    endpoint = endpoint.replace(
-      "%USER%",
-      Storage.get("settings", "INSTAGRAM_USER_ID"),
-    );
-    endpoint = endpoint.replace(
-      "%PAGE%",
-      Storage.get("settings", "INSTAGRAM_PAGE_ID"),
-    );
-
-    const url = new URL("https://graph.facebook.com");
-    url.pathname = this.GRAPH_API_VERSION + "/" + endpoint;
-    Logger.trace("POST", url.href);
-
-    return await fetch(url, {
-      method: "POST",
-      headers: {
-        Accept: "application/json",
-        Authorization:
-          "Bearer " + Storage.get("settings", "INSTAGRAM_PAGE_ACCESS_TOKEN"),
-      },
-      body: body,
-    })
-      .then((res) => this.handleApiResponse(res))
-      .catch((err) => this.handleApiError(err));
-  }
-
-  /**
-   * Handle api response
-   * @param response - the api response from fetch
-   * @returns the parsed response
-   */
-  private async handleApiResponse(response: Response): Promise<object> {
-    if (!response.ok) {
-      throw Logger.error("Ayrshare.handleApiResponse", response);
-    }
-    const data = await response.json();
-    if (data.error) {
-      const error =
-        response.status +
-        ":" +
-        data.error.type +
-        "(" +
-        data.error.code +
-        "/" +
-        data.error.error_subcode +
-        ") " +
-        data.error.message;
-      throw Logger.error("Facebook.handleApiResponse", error);
-    }
-    Logger.trace("Facebook.handleApiResponse", "success");
-    return data;
-  }
-
-  /**
-   * Handle api error
-   * @param error - the api error returned from fetch
-   */
-  private handleApiError(error: Error): never {
-    throw Logger.error("Facebook.handleApiError", error);
   }
 }
