+++ src/platforms/LinkedIn/LinkedInAuth.ts
@@ -1,8 +1,8 @@
-import Logger from "../services/Logger";
-import OAuth2Client from "./OAuth2Client";
-import Storage from "../services/Storage";
+import Logger from "../../services/Logger";
+import OAuth2Service from "../../services/OAuth2Service";
+import Storage from "../../services/Storage";
 
-export default class LinkedInAuth extends OAuth2Client {
+export default class LinkedInAuth {
   API_VERSION = "v2";
   accessToken = "";
 
@@ -58,7 +58,7 @@ export default class LinkedInAuth extends OAuth2Client {
     url.pathname = "oauth/" + this.API_VERSION + "/authorization";
     const query = {
       client_id: clientId,
-      redirect_uri: this.getCallbackUrl(),
+      redirect_uri: OAuth2Service.getCallbackUrl(),
       state: state,
       response_type: "code",
       duration: "permanent",
@@ -70,7 +70,10 @@ export default class LinkedInAuth extends OAuth2Client {
     };
     url.search = new URLSearchParams(query).toString();
 
-    const result = await this.requestRemotePermissions("LinkedIn", url.href);
+    const result = await OAuth2Service.requestRemotePermissions(
+      "LinkedIn",
+      url.href,
+    );
     if (result["error"]) {
       const msg = result["error_reason"] + " - " + result["error_description"];
       throw Logger.error(msg, result);
@@ -94,7 +97,7 @@ export default class LinkedInAuth extends OAuth2Client {
     refresh_token: string;
   }> {
     Logger.trace("RedditAuth", "exchangeCode", code);
-    const redirectUri = this.getCallbackUrl();
+    const redirectUri = OAuth2Service.getCallbackUrl();
 
     const result = (await this.post("accessToken", {
       grant_type: "authorization_code",
