+++ src/platforms/Twitter/Twitter.ts
@@ -1,15 +1,15 @@
 import * as fs from "fs";
 import * as sharp from "sharp";
 
-import Post, { PostStatus } from "../models/Post";
+import Post, { PostStatus } from "../../models/Post";
 
-import Folder from "../models/Folder";
-import Logger from "../services/Logger";
-import Platform from "../models/Platform";
-import { PlatformId } from ".";
-import Storage from "../services/Storage";
+import Folder from "../../models/Folder";
+import Logger from "../../services/Logger";
+import Platform from "../../models/Platform";
+import { PlatformId } from "..";
+import Storage from "../../services/Storage";
 import { TwitterApi } from "twitter-api-v2";
-import TwitterAuth from "../auth/TwitterAuth";
+import TwitterAuth from "./TwitterAuth";
 
 /**
  * Twitter: support for twitter platform
@@ -71,6 +71,7 @@ export default class Twitter extends Platform {
   }
 
   async publishPost(post: Post, dryrun: boolean = false): Promise<boolean> {
+    Logger.trace("Twitter.publishPost", post.id, dryrun);
     const client1 = new TwitterApi({
       appKey: Storage.get("settings", "TWITTER_OA1_API_KEY"),
       appSecret: Storage.get("settings", "TWITTER_OA1_API_KEY_SECRET"),
