+++ src/platforms/index.ts
@@ -1,8 +1,8 @@
-export { default as Facebook } from "./Facebook";
-export { default as Instagram } from "./Instagram";
-export { default as Twitter } from "./Twitter";
-export { default as Reddit } from "./Reddit";
-export { default as LinkedIn } from "./LinkedIn";
+export { default as Facebook } from "./Facebook/Facebook";
+export { default as Instagram } from "./Instagram/Instagram";
+export { default as Twitter } from "./Twitter/Twitter";
+export { default as Reddit } from "./Reddit/Reddit";
+export { default as LinkedIn } from "./LinkedIn/LinkedIn";
 export { default as AsYouTube } from "./Ayrshare/AsYouTube";
 export { default as AsInstagram } from "./Ayrshare/AsInstagram";
 export { default as AsTwitter } from "./Ayrshare/AsTwitter";
